move general dtos to the models module
