+++ models/src/main/java/edu/java/dto/bot/LinkUpdate.java
@@ -1,12 +1,15 @@
-package edu.java.bot.dto;
+package edu.java.dto.bot;
 
+import java.net.URI;
 import java.util.List;
+import lombok.Builder;
 import lombok.Data;
 
 @Data
+@Builder
 public class LinkUpdate {
     private Long id;
-    private String url;
+    private URI url;
     private String description;
     private List<Integer> tgChatIds;
 }
