+++ models/src/main/java/edu/java/dto/scrapper/request/RemoveLinkRequest.java
@@ -1,4 +1,4 @@
-package edu.java.dto;
+package edu.java.dto.scrapper.request;
 
 import lombok.Data;
