+++ models/src/main/java/edu/java/dto/scrapper/response/LinkResponse.java
@@ -1,0 +1,10 @@
+package edu.java.dto.scrapper.response;
+
+import java.net.URI;
+import lombok.Data;
+
+@Data
+public class LinkResponse {
+    private Long id;
+    private URI url;
+}
