+++ pom.xml
@@ -48,6 +48,7 @@
     <modules>
         <module>bot</module>
         <module>scrapper</module>
+        <module>models</module>
     </modules>
 
     <dependencyManagement>
